Client rotating it's own view working.
Client clicking a rotate button will successfully change the game state,
and return the new view to the client.  However, the server does not
update EVERY client's view that was effected by the game state change.
Must implement a way to keep track of who's view should change, and send
back an update at the end of game state processing.